+++ packages/winnow/lib/filter-and-transform/transforms/to-esri-attributes.js
@@ -1,22 +1,29 @@
+const _ = require('lodash')
 const { createIntegerHash } = require('../helpers')
 
-module.exports = function transformToEsriProperties (properties, geometry, dateFields, requiresObjectId, idField) {
+module.exports = function transformToEsriProperties (inputProperties, geometry, delimitedDateFields, requiresObjectId, idField) {
   requiresObjectId = requiresObjectId === 'true'
   idField = idField === 'null' ? null : idField
-  const transformedDateFields = transformDateFields(properties, dateFields)
-  const transformedProperties = { ...properties, ...transformedDateFields }
 
-  if (requiresObjectId) {
-    if (!idField) {
-      const OBJECTID = createIntegerHash(JSON.stringify({ properties: transformedProperties, geometry }))
-      return { ...transformedProperties, OBJECTID }
-    }
+  const dateFields = delimitedDateFields.split(',')
+  const properties = transformProperties(inputProperties, dateFields)
 
-    if (shouldLogIdFieldWarning(properties[idField])) {
-      console.warn(`WARNING: OBJECTIDs created from provider's "idField" (${idField}: ${properties[idField]}) are not integers from 0 to 2147483647`)
-    }
+  if (requiresObjectId && !idField) {
+    return injectObjectId({ properties, geometry })
+  } else if (requiresObjectId && shouldLogIdFieldWarning(properties[idField])) {
+    console.warn(`WARNING: OBJECTIDs created from provider's "idField" (${idField}: ${inputProperties[idField]}) are not integers from 0 to 2147483647`)
   }
-  return transformedProperties
+
+  return properties
+}
+
+function injectObjectId (feature) {
+  const { properties } = feature
+  const OBJECTID = createIntegerHash(JSON.stringify(feature))
+  return {
+    ...properties,
+    OBJECTID
+  }
 }
 
 function shouldLogIdFieldWarning (idField) {
@@ -25,13 +32,15 @@ function shouldLogIdFieldWarning (idField) {
     (!Number.isInteger(idField) || idField > 2147483647)
 }
 
-function transformDateFields (properties, delimitedDateFields) {
-  if (delimitedDateFields.length === 0) return properties
-  const dateFields = delimitedDateFields.split(',')
-
-  return dateFields.reduce((acc, field) => {
-    const value = properties[field]
-    acc[field] = value === null ? null : new Date(value).getTime()
-    return acc
+function transformProperties (properties, dateFields) {
+  return Object.entries(properties).reduce((transformedProperties, [key, value]) => {
+    if (dateFields.includes(key)) {
+      transformedProperties[key] = value === null ? null : new Date(value).getTime()
+    } else if (_.isObject(value)) {
+      transformedProperties[key] = JSON.stringify(value)
+    } else {
+      transformedProperties[key] = value
+    }
+    return transformedProperties
   }, {})
 }